fix: keep numpy integers out of the kernel metadata
With `align_shr_mem=0`, generating a kernel with a sparse operand failed in
`_write_kernel_meta_data`:

  TypeError: Object of type int64 is not JSON serializable

A sparse operand's stage size is a numpy count, and
`compute_shared_mem_size` passed it on unchanged.  It rode through the
region offsets into `ShrMemObject.set_size_per_mult` and from there into
`LaunchConfig.shared_elements`, `shared_bytes` and each section's
`shared_elements`, which `kernel_info()` hands to `json.dumps`.  With the
default alignment the rounding in `ShrMemAnalyzer._aligned` happened to
hide it.  Seen on sm_80 and sm_90 for gemm_sparse_band_B and its F64
sibling; every target is affected, since the value is target-independent.

Return an `int` from `compute_shared_mem_size`, where the size enters the
allocator.

Checked: gemm_sparse_band_B(_f64) now generates under align_shr_mem=0 on
sm_80, gfx90a and pvc, with an int shared size; test_packing,
test_kernel_metadata, test_storage_parts, test_launch_bounds,
test_sparse_layout pass.  test_staging's three
test_what_is_emitted_stays_under_what_was_planned cases fail before and
after this change alike.

AI-generated. Model: Opus 5

+++ src/tensorforge/backend/instructions/memory/__init__.py
@@ -305,7 +305,11 @@ class AbstractShrMemWrite(MemoryInstruction):
   def compute_shared_mem_size(self) -> int:
     # What the region allocator must reserve: every stage at once.  Returning
     # the per-stage size here would silently overlap the stages.
-    return self._stages * self.stage_size()
+    # `int`, because a sparse operand's stage size is a numpy count, and it
+    # rides through every offset into `LaunchConfig.shared_elements`, where
+    # `json.dumps` in the kernel metadata refuses it -- unless
+    # `align_shr_mem` happens to round it back into an `int` on the way.
+    return int(self._stages * self.stage_size())
 
   def set_shr_mem_offset(self, offset: int, first: bool, global_offset: bool) -> None:
     self._shr_mem_offset = offset
